Updated tests to work. Updated test names to remove trello. Added required var to .env.test Updated run-tests.yml to check CI/CD pipelines. Added mongomock lib to poetry for tests.
