+++ .env.test
@@ -10,3 +10,6 @@ SECRET_KEY=secret-key
 TRELLO_BOARD_ID=board-id
 TRELLO_API_KEY=api-key
 TRELLO_API_TOKEN=api-token
+
+DB_CONNECTION_STRING=mongodb://fakemongo.com
+DB_NAME_ITEMS=test_items
