+++ .github/workflows/run-tests.yml
@@ -20,7 +20,7 @@ jobs:
         name: Job Two
         runs-on: ubuntu-latest
         needs: job-one 
-        if: ${{ github.event_name == 'pull_request' && github.ref == 'refs/heads/main'}}
+        #if: ${{ github.event_name == 'pull_request' && github.ref == 'refs/heads/main'}}
         steps:
         - uses: actions/checkout@v2
         - run: echo "Publishing!"
