+++ pyproject.toml
@@ -12,6 +12,7 @@ requests = "^2.27.1"
 pytest = "^7.1.1"
 gunicorn = "^20.1.0"
 pymongo = "^4.3.3"
+mongomock = "^4.1.2"
 
 [tool.poetry.dev-dependencies]
